Reimplemented post survey questionaire with material ui
I reimplemented the post survey component with material ui.  I am still having an issue where I have to click twice on the last radio button to get the next button to disable.  There appears to be a delay between when the state function changes the value so that in checkEnableSubmit the button isn't enabled right after the last radio is clicked.