Got initial test working!
I had to change the package.json test script to fire up the emulator so the emulator starts and ends at the end of each test which is inconvenient.  --watch didn't seem to work on mocha.  Next step is to see if I can fire-up the emulator once and remove initializeTestEnvironment code (don't initialize emulator in code) and just initialize app.