Have writeState working with local emulator.
Switched back to synchronous state dispatches and all dispatches to redux now post the state to the google function which writes it to firestore.
Removed service methods that aren't used since we just save the redux state on every dispatch. No need for separate update service methods.  This is like how the writeStateCSV method worked, but it does it via the store.subscribe function asynchronously instead of within the dispatch (which was never really proper react).