File IO adapter functions no longer used in web app.
With the redux state being written to firestore, we no longer need to flatten the state to write to the S3 bucket so the FileIOAdapter.js functions aren't needed in the web app. We may need them for the CLI to extract the CSV file, so I moved them to that project.