Fixed dependencies (#141)
* Updated dependencies

* Fixed download:int target to run for empty folder and cleanup after download.

Fixed the target to work when the import folder is empty and move the downloaded database export files up one level and delete the subdirectory.

* Removed craco and updated yarn

* Converted to next js.

* Completed conversion to nextjs

Screens are working for between subject study.  There is an issue where it appears as though the study explanation screen reloads before advancing to the finished screen.